Update hoNDArray_reductions.h
Fixed the description of "dotc" to match the actual computation.
I'm not sure if the comment was wrong or the code is wrong, but I didn't want to change the code because there are some files using dotc and I didn't want to change the behavior.  If the intent of the code is what the previous description said, then the code is wrong and should be changed.