COMP: Prefer implicit CMake mechanisms for ISMRMRD
ISMRMRD has been instrumented to provide implicit cmake
compatible package definitions.  This can make long
term maintenance of joining these packages together
easier.  ISMRMRD is now responsible for defining
the cmake elements that are available to external
projects.

In addition, the naming of the ISMRMRD package to
be included with find_package was changed to all
capital letters in order to more closely conform
to a consistent naming convention at the package
level.

Finally, there is no need to repeatedly include
find_package(ISMRMRD) in sub-directories.  The
single top level CMakeLists.txt call is sufficent
for the entire package.