Remove webapp
The webapp (written in python) has been removed. It offered no functionality not offered by projects such as supervisord (http://supervisord.org/). Supervisor is already being used in the Docker containers. It is an unnecessary complication to use supervisord to start the webapp when supervisord can be configured to do the same.