Add gadgets, toolboxes and core build directories into rpath
On Linux, when an application needs to load a shared library, it will search
in these directories in this order:
1.Directories listed in the executable’s rpath.
2.Directories in the LD_LIBRARY_PATH environment variable, which contains colon-separated list of directories (e.g., /path/to/libdir:/another/path)
3.Directories listed in the executable’s runpath.
4.The list of directories in the file /etc/ld.so.conf. This file can include other files, but it is basically a list of directories - one per line.
5.Default system libraries - usually /lib and /usr/lib (skipped if compiled with -z nodefaultlib).

Before this commit LD_LIBRARY_PATH environment variable must be
modified before launching gadgetron to load the readers, writers,
gadgets, and toolboxes shared libraries.
Modifying this variable can have side effects on other applications
which will load gadgetron shared libraries instead of the suitable ones.

So this is why we prefer to modify rpath of the gadgetron apps to start
gadgetron without modifying any environment variable.