Package Gadgetron in Gadgetron
Historically the Gadgetron conda package was built as part of:

https://github.com/gadgetron/gadgetron-conda-recipes

However this build process took a very long time and coupled everything.

This PR brings over the packaging of Gadgetron from gadgetron-conda-recipes
into the Gadgetron repo.

The package will always be built, but published on succesful CI builds.
The version is determined automagically from the root CMakeLists.txt which
gives us a single source of truth for versioning.

In addition support for validating the package was achieved by adding the
Gadgetron integration tests to the packaging step via run_test.sh. This
will catch things such as missing runtime dependencies in the package meta.

Included in the changes is a bug revision uprev to signify a new package
being built, and a timeout increase to get_data.py (it was insufficient).