Accept package revisions when building package publish environment (#1134)
The create command was asking for stdin (Y/N). The command optionally takes in a -y command to always say yes. 

Also fixed the exception handling (typo) I added in a prior commit. During data downloads an exception was propagating out the main instead of being handled by the retry logic in get_data.py. Since this is likely not intended I added the additional case but copied parts of the string representation of the exception which was incorrect (as the exception was obtained from the pipeline logs).