Support reconstructions from local data  (#1158)
These changes add support for reading raw gadgetron message streams either
from a file or taken through stdin via an unnamed pipe as per:

https://github.com/gadgetron/gadgetron/issues/1126

From a high level this involves the creation of adapters (mrd_to_stream and stream_to_mrd)
responsible for serialising and deserialising from ISMRMRD to a gadgetron stream for
consumption within the gadgetron and back. This enables a user to chain gadgetron instances
to perform multiple reconstructions in sequence, which provides flexibility as the output
from the gadgetron can be written to a file and consumed later.

    | mrd |--->| input_adapter |--->| gadgetron |--->| output_adapter |--->| mrd |

Included are example stream xml configurations which perform post processing, 
and two tests that leverage them which chain reconstructions via unnamed pipes.

Usage is simple:

    ismrmrd_generate_cartesian_shepp_logan
    mrd_to_stream -i testdata.h5 | gadgetron --from_stream -c default.xml | stream_to_mrd -o out.h5

But can be more complicated:

    ismrmrd_generate_cartesian_shepp_logan
    mrd_to_stream -i testdata.h5 | gadgetron --from_stream -c Generic_Cartesian_Grappa_Complex.xml | gadgetron --from_stream -c stream_complex_to_float.xml | gadgetron --from_stream -c stream_float_to_short.xml |  stream_to_mrd -o out.h5

Also output can be redirected to a file and consumed later enabling post processing.

    ismrmrd_generate_cartesian_shepp_logan
    mrd_to_stream -i testdata.h5 | gadgetron --from_stream -c Generic_Cartesian_Grappa_Complex.xml -o tmp.dat
    gadgetron --from_stream -c stream_complex_to_float.xml -i tmp.dat | gadgetron --from_stream -c stream_float_to_short.xml |  stream_to_mrd -o out1.h5
    or
    cat out.dat | gadgetron --from_stream -c stream_complex_to_float.xml | gadgetron --from_stream -c stream_float_to_short.xml |  stream_to_mrd -o out2.h5

These changes also modify the logging macros to write to stderr, and replace
all instances of std::cout with macro usage. The test infrastructure will now validate that stdout is
empty.