AcquisitionAccumulateTriggerGadget: add trigger for parallel_calibration scans
External reference scans are usually acquired at the beginning of a sequence.
Thus, to save total reconstruction time, it makes sense to trigger the coil
sensitivity calibration directly after the reference scan is done, and not only
after the first full image acquisition is complete. This is especially the case
for multi-minute high-resolution scans in combination with a computationally
demanding coil calibration technique. This PR adds a "parallel_calibration"
trigger option that fires just after the IS_PARALLEL_CALIBRATION flag is removed
(presumably at the beginning of the imaging scans).